+++ internal/cloud/aws_test.go
@@ -167,6 +167,7 @@ func TestAWSCloud_Send(t *testing.T) {
 				CreatedAt: time.Date(2016, 12, 27, 8, 14, 53, 0, time.UTC),
 				Checksum:  "cb63324d2c35cdfcb4521e15ca4518bd0ed9dc2364a9f47de75151b3f9b4b705",
 				VaultName: "vault",
+				Size:      41,
 			},
 		},
 		{
@@ -315,6 +316,7 @@ func TestAWSCloud_Send(t *testing.T) {
 				CreatedAt: time.Date(2016, 12, 27, 8, 14, 53, 0, time.UTC),
 				Checksum:  "a6d392677577af12fb1f4ceb510940374c3378455a1485b0226a35ef5ad65242",
 				VaultName: "vault",
+				Size:      42000,
 			},
 		},
 		{
@@ -752,12 +754,14 @@ func TestAWSCloud_List(t *testing.T) {
 					CreatedAt: time.Date(2016, 11, 7, 12, 0, 0, 0, time.UTC),
 					Checksum:  "223072246f6eedbf1271bd1576f01b4b67c8e1cb1142599d5ef615673f513a5f",
 					VaultName: "vault",
+					Size:      2456,
 				},
 				{
 					ID:        "AWSID123",
 					CreatedAt: time.Date(2016, 12, 27, 8, 14, 53, 0, time.UTC),
 					Checksum:  "a75e723eaf6da1db780e0a9b6a2046eba1a6bc20e8e69ffcb7c633e5e51f2502",
 					VaultName: "vault",
+					Size:      4000,
 				},
 			},
 		},
@@ -949,6 +953,7 @@ func TestAWSCloud_List(t *testing.T) {
 					CreatedAt: time.Date(2016, 12, 27, 8, 14, 53, 0, time.UTC),
 					Checksum:  "a75e723eaf6da1db780e0a9b6a2046eba1a6bc20e8e69ffcb7c633e5e51f2502",
 					VaultName: "vault",
+					Size:      4000,
 				},
 			},
 		},