+++ internal/storage/audit_file_test.go
@@ -7,6 +7,7 @@ import (
 	"os"
 	"path"
 	"reflect"
+	"strconv"
 	"strings"
 	"testing"
 	"time"
@@ -51,8 +52,9 @@ func TestAuditFile_Save(t *testing.T) {
 				CreatedAt: now,
 				Checksum:  "ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7",
 				VaultName: "test",
+				Size:      120,
 			},
-			expected: fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7\n", now.Format(time.RFC3339)),
+			expected: fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 120\n", now.Format(time.RFC3339)),
 		},
 		{
 			description: "it should detect when the filename refers to a directory",
@@ -132,6 +134,73 @@ func TestAuditFile_List(t *testing.T) {
 				}
 				defer f.Close()
 
+				f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 120\n", now.Format(time.RFC3339)))
+				return f.Name()
+			}(),
+			expected: []cloud.Backup{
+				{
+					ID: "123456",
+					CreatedAt: func() time.Time {
+						c, err := time.Parse(time.RFC3339, now.Format(time.RFC3339))
+						if err != nil {
+							t.Fatalf("error parsing current time. details: %s", err)
+						}
+						return c
+					}(),
+					Checksum:  "ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7",
+					VaultName: "test",
+					Size:      120,
+				},
+			},
+		},
+		{
+			description: "it should list all backups information correctly with format transition",
+			logger: mockLogger{
+				mockDebug:  func(args ...interface{}) {},
+				mockDebugf: func(format string, args ...interface{}) {},
+				mockInfo:   func(args ...interface{}) {},
+				mockInfof:  func(format string, args ...interface{}) {},
+			},
+			filename: func() string {
+				f, err := ioutil.TempFile("", "toglacier-test")
+				if err != nil {
+					t.Fatalf("error creating a temporary file. details: %s", err)
+				}
+				defer f.Close()
+
+				f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 0  \n", now.Format(time.RFC3339)))
+				return f.Name()
+			}(),
+			expected: []cloud.Backup{
+				{
+					ID: "123456",
+					CreatedAt: func() time.Time {
+						c, err := time.Parse(time.RFC3339, now.Format(time.RFC3339))
+						if err != nil {
+							t.Fatalf("error parsing current time. details: %s", err)
+						}
+						return c
+					}(),
+					Checksum:  "ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7",
+					VaultName: "test",
+				},
+			},
+		},
+		{
+			description: "it should list all backups information correctly with backward compatibility",
+			logger: mockLogger{
+				mockDebug:  func(args ...interface{}) {},
+				mockDebugf: func(format string, args ...interface{}) {},
+				mockInfo:   func(args ...interface{}) {},
+				mockInfof:  func(format string, args ...interface{}) {},
+			},
+			filename: func() string {
+				f, err := ioutil.TempFile("", "toglacier-test")
+				if err != nil {
+					t.Fatalf("error creating a temporary file. details: %s", err)
+				}
+				defer f.Close()
+
 				f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7\n", now.Format(time.RFC3339)))
 				return f.Name()
 			}(),
@@ -179,7 +248,7 @@ func TestAuditFile_List(t *testing.T) {
 					}
 					defer f.Close()
 
-					f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7\n", now.Format(time.RFC3339)))
+					f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 120\n", now.Format(time.RFC3339)))
 				}
 
 				return n
@@ -254,7 +323,7 @@ func TestAuditFile_List(t *testing.T) {
 				}
 				defer f.Close()
 
-				f.WriteString("XXXX test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7\n")
+				f.WriteString("XXXX test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 120\n")
 				return f.Name()
 			}(),
 			expectedError: &storage.Error{
@@ -265,6 +334,33 @@ func TestAuditFile_List(t *testing.T) {
 					LayoutElem: "2006",
 					ValueElem:  "XXXX",
 					Message:    fmt.Sprintf(` as "%s": cannot parse "XXXX" as "2006"`, time.RFC3339),
+				},
+			},
+		},
+		{
+			description: "it should detect when the audit file contains an invalid size",
+			logger: mockLogger{
+				mockDebug:  func(args ...interface{}) {},
+				mockDebugf: func(format string, args ...interface{}) {},
+				mockInfo:   func(args ...interface{}) {},
+				mockInfof:  func(format string, args ...interface{}) {},
+			},
+			filename: func() string {
+				f, err := ioutil.TempFile("", "toglacier-test")
+				if err != nil {
+					t.Fatalf("error creating a temporary file. details: %s", err)
+				}
+				defer f.Close()
+
+				f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 XXXX\n", now.Format(time.RFC3339)))
+				return f.Name()
+			}(),
+			expectedError: &storage.Error{
+				Code: storage.ErrorCodeSizeFormat,
+				Err: &strconv.NumError{
+					Func: "ParseInt",
+					Num:  "XXXX",
+					Err:  errors.New("invalid syntax"),
 				},
 			},
 		},
@@ -312,12 +408,34 @@ func TestAuditFile_Remove(t *testing.T) {
 				}
 				defer f.Close()
 
+				f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 100\n", now.Format(time.RFC3339)))
+				f.WriteString(fmt.Sprintf("%s test 123457 913b87897ffb6dca07e9f17e280aa8ecb9886dffeda8a15efeafec11dec0d108 200\n", now.Add(time.Second).Format(time.RFC3339)))
+				return f.Name()
+			}(),
+			id:       "123457",
+			expected: fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 100\n", now.Format(time.RFC3339)),
+		},
+		{
+			description: "it should remove a backup information correctly with backward compatibility",
+			logger: mockLogger{
+				mockDebug:  func(args ...interface{}) {},
+				mockDebugf: func(format string, args ...interface{}) {},
+				mockInfo:   func(args ...interface{}) {},
+				mockInfof:  func(format string, args ...interface{}) {},
+			},
+			filename: func() string {
+				f, err := ioutil.TempFile("", "toglacier-test")
+				if err != nil {
+					t.Fatalf("error creating a temporary file. details: %s", err)
+				}
+				defer f.Close()
+
 				f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7\n", now.Format(time.RFC3339)))
 				f.WriteString(fmt.Sprintf("%s test 123457 913b87897ffb6dca07e9f17e280aa8ecb9886dffeda8a15efeafec11dec0d108\n", now.Add(time.Second).Format(time.RFC3339)))
 				return f.Name()
 			}(),
 			id:       "123457",
-			expected: fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7\n", now.Format(time.RFC3339)),
+			expected: fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 0\n", now.Format(time.RFC3339)),
 		},
 		{
 			description: "it should detect when the audit file has no read permission",